MGS-3150 [#ccc] Integrate 6.2.3 official release
Integrated more bug-fixing for i.MX8 GPUs bring-ups,
CTS passed on i.MX8MScale and i.MX8QM, ES31 CTS failed on i.MX8QXP.

Upgrade OpenVX1.1 driver, 255 CTS known issues in new features,
This release will target for i.MX8 Alpha BSP releases.

Date: Aug 01, 2017

+++ driver/hal/os/qnx/kernel/gc_hal_kernel_device.c
@@ -751,7 +751,7 @@ gckGALDEVICE_Construct(
             {
                 gcmkTRACE_ZONE(gcvLEVEL_INFO, gcvZONE_DRIVER,
                         "[galcore] gckGALDEVICE_Construct: Unable to map location->0x%lX for size->%ld\n",
-                        device->requestedRegisterMemBases[i],
+                        (unsigned long)device->requestedRegisterMemBases[i],
                         device->requestedRegisterMemSizes[i]);
 
                 return gcvSTATUS_OUT_OF_RESOURCES;
